requirements.txt
allure-pytest==2.13.5
playwright==1.44.0
pytest==8.2.1
python-dotenv==1.0.1
Faker
